backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pinned for passlib compatibility
email-validator

# WebSocket
websockets==12.0

# External APIs
httpx==0.25.2
aiohttp==3.9.3

# AI
google-generativeai==0.4.0

# Telegram
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
